+++ app/Codebelt.Extensions.Xunit.Hosting.BootstrapperMinimalConsole.App/BootstrapperMinimalConsoleMarker.cs
@@ -1,0 +1,11 @@
+namespace Codebelt.Extensions.Xunit.Hosting.BootstrapperMinimalConsole.App;
+
+public sealed class BootstrapperMinimalConsoleMarker
+{
+    public BootstrapperMinimalConsoleMarker(string value)
+    {
+        Value = value;
+    }
+
+    public string Value { get; }
+}