+++ app/Codebelt.Extensions.Xunit.Hosting.BootstrapperMinimalWorker.App/BootstrapperMinimalWorkerMarker.cs
@@ -1,0 +1,11 @@
+namespace Codebelt.Extensions.Xunit.Hosting.BootstrapperMinimalWorker.App;
+
+public sealed class BootstrapperMinimalWorkerMarker
+{
+    public BootstrapperMinimalWorkerMarker(string value)
+    {
+        Value = value;
+    }
+
+    public string Value { get; }
+}