+++ app/Codebelt.Extensions.Xunit.Hosting.BootstrapperWeb.App/BootstrapperWebMarker.cs
@@ -1,0 +1,11 @@
+namespace Codebelt.Extensions.Xunit.Hosting.BootstrapperWeb.App;
+
+public sealed class BootstrapperWebMarker
+{
+    public BootstrapperWebMarker(string value)
+    {
+        Value = value;
+    }
+
+    public string Value { get; }
+}