+++ app/Codebelt.Extensions.Xunit.Hosting.BootstrapperWorker.App/BootstrapperWorkerMarker.cs
@@ -1,0 +1,11 @@
+namespace Codebelt.Extensions.Xunit.Hosting.BootstrapperWorker.App;
+
+public sealed class BootstrapperWorkerMarker
+{
+    public BootstrapperWorkerMarker(string value)
+    {
+        Value = value;
+    }
+
+    public string Value { get; }
+}