+++ app/Codebelt.Extensions.Xunit.Hosting.ClassicProgram.App/ClassicProgramMarker.cs
@@ -1,0 +1,11 @@
+namespace Codebelt.Extensions.Xunit.Hosting.ClassicProgram.App;
+
+public sealed class ClassicProgramMarker
+{
+    public ClassicProgramMarker(string value)
+    {
+        Value = value;
+    }
+
+    public string Value { get; }
+}